Avoid creation of unnecessary nodes

+++ GopSolverLib/GopArray.h
@@ -2,8 +2,8 @@
 
 #include "Point.h"
 
-const int GRID_SIZE = 53;
-const int GRID_MAX = GRID_SIZE / 2;
+const int8_t GRID_SIZE = 53;
+const int8_t GRID_MAX = GRID_SIZE / 2;
 
 template<class T>
 class GopArray
@@ -12,8 +12,8 @@ public:
 	GopArray() { clear(); }
 
 	T& operator[](Point p) { return data[GRID_MAX - p.y][GRID_MAX + p.x]; }
-	T get(int x, int y) const { return data[GRID_MAX - y][GRID_MAX + x]; }
-	void set(int x, int y, T value) { data[GRID_MAX - y][GRID_MAX + x] = value; }
+	T get(int8_t x, int8_t y) const { return data[GRID_MAX - y][GRID_MAX + x]; }
+	void set(int8_t x, int8_t y, T value) { data[GRID_MAX - y][GRID_MAX + x] = value; }
 	void clear()
 	{
 		for (int y = 0; y < GRID_SIZE; ++y)
